make preview a typescript file
