+++ turbo.json
@@ -86,6 +86,7 @@
     "BOT_PROD_LOG_LEVEL",
     "MAXIMUM_CHANNEL_MESSAGES_PER_INDEX",
     "STORYBOOK",
-    "VITEST_DISCORD_CLIENT_ID"
+    "VITEST_DISCORD_CLIENT_ID",
+    "THEME"
   ]
 }
